支持自动判断会话 Cookie secure 标志
